Create library service
