+++ packages/cardhost/app/styles/components/cardhost-left-edge.css
@@ -9,14 +9,15 @@
 }
 
 .cardhost-left-edge--nav {
-  background-color: var(--ch-deep-background);
-  width: var(--ch-left-edge-navbar-width);
-  transition: width 0.2s ease;
-  overflow: hidden;
-}
-.cardhost-left-edge--nav.index-page,
-.cardhost-left-edge--nav.edit {
   background-color: var(--ch-default);
+  width: var(--ch-left-edge-navbar-width);
+  overflow: hidden;
+  transition: width 0.2s ease, var(--ch-card-background-color-animate);
+}
+
+.cardhost-left-edge--nav.schema {
+  background-color: var(--ch-dark-background);
+  border-right: 2px solid var(--ch-default);
 }
 
 .cardhost-left-edge--nav > ul {
@@ -70,8 +71,15 @@
   border-color: var(--ch-border);
 }
 
-.cardhost-left-edge--nav-button.cardstack-logo-button {
-  background-color: var(--ch-default);
+.cardhost-left-edge--nav-button.selected {
+  --icon-color: var(--ch-dark);
+  background-color: var(--ch-highlight);
+  border-color: var(--ch-border);
+}
+
+.cardhost-left-edge--nav-button.selected:hover,
+.cardhost-left-edge--nav-button.selected:focus {
+  background-color: var(--ch-highlight-hover);
 }
 
 .cardhost-left-edge--nav-button:disabled {
