+++ packages/cardhost/app/templates/cards.hbs
@@ -8,6 +8,13 @@
   {{this.themerClasses}}"
   data-test-cardhost-cards
 >
-  <CardhostTopEdge @mode={{this.routeInfo.mode}} @card={{this.target.currentRoute.parent.attributes}} />
+  {{#unless (eq this.routeInfo.mode "cards")}}
+    <CardhostTopEdge @mode={{this.routeInfo.mode}} @card={{this.target.currentRoute.parent.attributes}} />
+  {{/unless}}
+
   {{outlet}}
 </div>
+
+{{#if this.library.visible}}
+  {{in-bottom-toolbar show=(component "library" model=this.model)}}
+{{/if}}
