+++ packages/cardhost/app/templates/cards/index.hbs
@@ -1,0 +1,20 @@
+<CardhostLeftEdge />
+
+<main class="card-builder" data-test-card-builder>
+  <h1 class="card-builder--header">Card Builder</h1>
+  <section class="card-builder--section">
+    <h2 class="card-builder--section-header">
+      {{svg-jar "cards" width="28px" height="28px"}}
+      Featured Cards
+    </h2>
+    <div class="card-builder--section-grid">
+      {{#each this.featuredCards as |card|}}
+        <LinkTo @route="cards.card.view" @model={{card.id}} class="card-builder--section-card-link" data-test-featured-card={{card.id}}>
+          {{! template-lint-disable no-inline-styles }}
+          <div class="card-builder--section-card" style={{css-url "background-image" card.preview}}></div>
+          <div class="card-builder--section-card-label">{{card.title}}</div>
+        </LinkTo>
+      {{/each}}
+    </div>
+  </section>
+</main>
